requirements.txt
torch>=2.0.0
transformers>=4.30.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
PyYAML>=6.0
tqdm>=4.64.0
sentencepiece>=0.1.99
iterative-stratification
protobuf
captum>=0.6.0
google-generativeai>=0.7.0
diskcache>=5.6.0
